Adicionado as permissões de acesso parte 1

+++ app/Http/Controllers/Localizacao2Controller.php
@@ -10,6 +10,7 @@ use App\Cidade;
 use App\Praca;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Gate;
 
 class Localizacao2Controller extends Controller
 {
@@ -20,6 +21,14 @@ class Localizacao2Controller extends Controller
      */
     public function index(Request $request)
     {
+        //Autenticação caso não tenha permissão para visualizar os cadastros  
+        if(Gate::denies('cadastro-view')){
+            abort(403,"Não autorizado!");
+        }
+        if(Gate::denies('localizacao-view')){
+            abort(403,"Não autorizado!");
+        }
+
         $searchText=trim($request->get('searchText'));
        
         $localizacoes = Localizacao2::where('praca_id', '=', auth()->user()->praca->id)
@@ -54,6 +63,14 @@ class Localizacao2Controller extends Controller
      */
     public function create()
     {
+        //Autenticação caso não tenha permissão para visualizar os cadastros  
+        if(Gate::denies('cadastro-view')){
+            abort(403,"Não autorizado!");
+        }
+        if(Gate::denies('localizacao-create')){
+            abort(403,"Não autorizado!");
+        }
+
         $estados = Estado::orderBy("nome","ASC")->get();
         $cidades = Cidade::where('id', '=', 0)->get();
         $localizacoes1 = Localizacao1::where('id', '=', 0)->get();              
@@ -75,7 +92,15 @@ class Localizacao2Controller extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {       
+    {     
+        //Autenticação caso não tenha permissão para visualizar os cadastros  
+        if(Gate::denies('cadastro-view')){
+            abort(403,"Não autorizado!");
+        }
+        if(Gate::denies('localizacao-create')){
+            abort(403,"Não autorizado!");
+        }
+
         $validator = Validator::make($request->all(), [
             'estado_id' => [
                 'required',
@@ -129,7 +154,15 @@ class Localizacao2Controller extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {        
+    {      
+        //Autenticação caso não tenha permissão para visualizar os cadastros  
+        if(Gate::denies('cadastro-view')){
+            abort(403,"Não autorizado!");
+        }
+        if(Gate::denies('localizacao-edit')){
+            abort(403,"Não autorizado!");
+        }
+
         $estados = Estado::orderBy("nome","ASC")->get();
         $cidades = Cidade::orderBy("nome","ASC")->get();
         $localizacoes1 = Localizacao1::withTrashed()->where('praca_id', '=', auth()->user()->praca->id)->orderBy("nome","ASC")->get();       
@@ -153,7 +186,15 @@ class Localizacao2Controller extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
-    {      
+    {     
+        //Autenticação caso não tenha permissão para visualizar os cadastros  
+        if(Gate::denies('cadastro-view')){
+            abort(403,"Não autorizado!");
+        }
+        if(Gate::denies('localizacao-edit')){
+            abort(403,"Não autorizado!");
+        }
+
         $validator = Validator::make($request->all(), [
             'estado_id' => [
                 'required',
@@ -197,18 +238,16 @@ class Localizacao2Controller extends Controller
      */
     public function destroy($id)
     {
-        //if(Localizacao2::has('localizacao_equipamentos')->find($id) == null)
-        //{
-            $localizacao = Localizacao2::find($id);
+        //Autenticação caso não tenha permissão para visualizar os cadastros  
+        if(Gate::denies('cadastro-view')){
+            abort(403,"Não autorizado!");
+        }
+        if(Gate::denies('localizacao-delete')){
+            abort(403,"Não autorizado!");
+        }
 
-            $localizacao->delete();
-        /*}
-        else
-        {
-            return response()->json([
-                'fail' => true                
-            ]);
-        }*/
+        $localizacao = Localizacao2::find($id);
+        $localizacao->delete();      
     }
 
     public function getLocalizacao3($id)
